requirements.txt
antlr4-python3-runtime~=4.13.2
dtgraph @ git+https://github.com/yannramusat/DTGraph.git@main
neo4j~=5.28.1
pytest~=9.0.0
testcontainers~=4.13.3
pandas~=2.3.0
streamlit~=1.51.0
PyYAML~=6.0.2
ipython~=9.3.0
plotnine~=0.15.1
functional-dependencies~=1.3.0
case-converter
sphinx
sphinx-mdinclude
sphinx-book-theme
sphinx-paramlinks
pyyaml
pandas
clingo~=5.8.0